adds custom parameters for loadbalancer
Adds the ability to send custom parameters to the external
loadbalancer. When fetching the status is adds the custom
parameters to the query string. E.g. if the custom parameter
is called host and has a value test then then the url will be:

http://theloadbalancerurl/server-status?name=test&host=test

The new option is called LoadBalancerParameters and is nested
under the Agents node.

"Agents": {
	testagent1": { "LoadBalancerParameters": { "Host": "testgroup1" } },